added parameter for optional cleanup process

+++ setup/kafka_setup/kafka_setup.sh
@@ -1,5 +1,16 @@
 #! /bin/bash
 
+# get value of optional parameter with-cleanup
+cleanup=false
+while [ "$1" ]; do
+        case $1 in
+            --with-cleanup) cleanup=true ;;
+            *) echo 'Error in command line parsing' >&2
+               exit 1
+        esac
+        shift
+    done
+
 # color codes for better terminal output
 GREEN='\033[0;32m'  # green color
 NC='\033[0m' # No Color
@@ -8,27 +19,31 @@ NC='\033[0m' # No Color
 echo -e "${GREEN}Creating Docker Containers...${NC}"
 docker-compose up -d
 
+# wait 15 seconds until kafka is fully functional to avoid warnings
+sleep 15
+
 # create topics in kafka container
 echo -e "${GREEN}Creating Topics in Kafka Container...${NC}"
-# docker exec -it kafka /bin/bash
-docker exec -it kafka /opt/kafka_*/bin/kafka-topics.sh --create --zookeeper zookeeper:2181 --replication-factor 1 --partitions 1 --topic TH1
 
-docker exec -it kafka /opt/kafka_*/bin/kafka-topics.sh --create --zookeeper zookeeper:2181 --replication-factor 1 --partitions 1 --topic HVAC1
+docker exec -it kafka kafka-topics.sh --create  --if-not-exists --bootstrap-server localhost:9092 --replication-factor 1 --partitions 1 --topic TH1
 
-docker exec -it kafka /opt/kafka_*/bin/kafka-topics.sh --create --zookeeper zookeeper:2181 --replication-factor 1 --partitions 1 --topic Etot
+docker exec -it kafka kafka-topics.sh --create --if-not-exists --bootstrap-server localhost:9092 --replication-factor 1 --partitions 1 --topic HVAC1
 
-docker exec -it kafka /opt/kafka_*/bin/kafka-topics.sh --create --zookeeper zookeeper:2181 --replication-factor 1 --partitions 1 --topic W1
+docker exec -it kafka kafka-topics.sh --create --if-not-exists --bootstrap-server localhost:9092 --replication-factor 1 --partitions 1 --topic Etot
 
+docker exec -it kafka kafka-topics.sh --create --if-not-exists --bootstrap-server localhost:9092 --replication-factor 1 --partitions 1 --topic W1
 
 # docker clean-up
-echo -e "${GREEN}Clean up Process...${NC}"
-docker stop kafka   # stop kafka container
-docker rm kafka     # remove kafka container
-docker stop zookeeper # stop zookeeper container
-docker rm zookeeper   # remove zookeeper container
+if "$cleanup"; then
+    echo -e "${GREEN}Clean up Process...${NC}"
+    docker stop kafka   # stop kafka container
+    docker rm kafka     # remove kafka container
+    docker stop zookeeper # stop zookeeper container
+    docker rm zookeeper   # remove zookeeper container
 
-docker image rm confluentinc/cp-kafka:latest     # remove kafka image
-docker image rm confluentinc/cp-zookeeper:latest  # remove zookeeper image
+    docker image rm bitnami/kafka:latest     # remove kafka image
+    docker image rm bitnami/zookeeper:latest # remove zookeeper image
 
-docker volume prune    # remove unused volumes
+    docker volume prune -f   # remove unused volumes
+fi
